feat: migrate domain models to the domain crate
