+++ crates/actors/src/block_producer.rs
@@ -1,9 +1,5 @@
 use crate::{
     block_discovery::{BlockDiscoveredMessage, BlockDiscoveryActor},
-    block_tree_service::{
-        ema_snapshot::{EmaSnapshot, ExponentialMarketAvgCalculation},
-        BlockTreeReadGuard,
-    },
     broadcast_mining_service::{BroadcastDifficultyUpdate, BroadcastMiningService},
     mempool_service::MempoolServiceMessage,
     reth_service::{BlockHashType, ForkChoiceUpdateMessage, RethServiceActor},
@@ -22,7 +18,10 @@ use alloy_signer_local::LocalSigner;
 use base58::ToBase58 as _;
 use eyre::eyre;
 use irys_database::{block_header_by_hash, db::IrysDatabaseExt as _, SystemLedger};
-use irys_domain::block_index_guard::BlockIndexReadGuard;
+use irys_domain::{
+    block_index_guard::BlockIndexReadGuard, BlockTreeReadGuard, EmaSnapshot,
+    ExponentialMarketAvgCalculation,
+};
 use irys_price_oracle::IrysPriceOracle;
 use irys_reth::compose_shadow_tx;
 use irys_reth_node_bridge::IrysRethNodeAdapter;
