+++ crates/chain/tests/api/tx_commitments.rs
@@ -4,8 +4,7 @@ use base58::ToBase58 as _;
 use eyre::eyre;
 use irys_actors::packing::wait_for_packing;
 use irys_chain::IrysNodeCtx;
-use irys_database::CommitmentSnapshotStatus;
-use irys_domain::EpochSnapshot;
+use irys_domain::{CommitmentSnapshotStatus, EpochSnapshot};
 use irys_primitives::CommitmentType;
 use irys_testing_utils::initialize_tracing;
 use irys_types::{irys::IrysSigner, Address, CommitmentTransaction, NodeConfig, H256};
