+++ crates/database/src/lib.rs
@@ -1,8 +1,6 @@
 //! This crate is a dependency for both [chain] and [actors] crates. It exposes
 //! database methods for reading and writing from the database as well as some
 //! database value types.
-pub mod block_index_data;
-pub mod commitment_snapshot;
 pub mod config;
 pub mod data_ledger;
 pub mod database;
@@ -26,8 +24,6 @@ pub mod submodule;
 /// Local macro definition of chain specific mdbx tables
 pub mod tables;
 
-pub use block_index_data::*;
-pub use commitment_snapshot::*;
 pub use data_ledger::*;
 pub use database::*;
 pub use system_ledger::*;
