+++ crates/domain/src/guards/block_tree_guard.rs
@@ -1,0 +1,26 @@
+use crate::BlockTree;
+use std::sync::{Arc, RwLock, RwLockReadGuard};
+
+/// Wraps the internal `Arc<RwLock<_>>` to make the reference readonly
+#[derive(Debug, Clone)]
+pub struct BlockTreeReadGuard {
+    block_tree_cache: Arc<RwLock<BlockTree>>,
+}
+
+impl BlockTreeReadGuard {
+    /// Creates a new `ReadGuard` for the `block_tree` cache
+    pub const fn new(block_tree_cache: Arc<RwLock<BlockTree>>) -> Self {
+        Self { block_tree_cache }
+    }
+
+    /// Accessor method to get a read guard for the `block_tree` cache
+    pub fn read(&self) -> RwLockReadGuard<'_, BlockTree> {
+        self.block_tree_cache.read().unwrap()
+    }
+
+    #[cfg(any(test, feature = "test-utils"))]
+    /// Accessor method to get a write guard for the `block_tree` cache
+    pub fn write(&self) -> std::sync::RwLockWriteGuard<'_, BlockTree> {
+        self.block_tree_cache.write().unwrap()
+    }
+}
