+++ crates/domain/src/guards/mod.rs
@@ -1,2 +1,4 @@
 pub mod block_index_guard;
+pub mod block_tree_guard;
 pub use block_index_guard::*;
+pub use block_tree_guard::*;
