+++ crates/domain/src/lib.rs
@@ -1,5 +1,9 @@
-pub mod epoch_snapshot;
+pub mod domain_test_utils;
 pub mod guards;
+pub mod models;
+pub mod snapshots;
 
-pub use epoch_snapshot::*;
+pub use domain_test_utils::*;
 pub use guards::*;
+pub use models::*;
+pub use snapshots::*;
