+++ crates/domain/src/models/block_index.rs
