+++ crates/domain/src/models/mod.rs
@@ -1,0 +1,5 @@
+pub mod block_index;
+pub mod block_tree;
+
+pub use block_index::*;
+pub use block_tree::*;
