+++ crates/domain/src/snapshots/commitment_snapshot.rs
