+++ crates/domain/src/snapshots/ema_snapshot.rs
