+++ crates/domain/src/snapshots/epoch_snapshot/commitment_state.rs
