+++ crates/domain/src/snapshots/epoch_snapshot/epoch_snapshot.rs
