+++ crates/domain/src/snapshots/epoch_snapshot/partition_assignments.rs
