+++ crates/domain/src/snapshots/mod.rs
@@ -1,0 +1,7 @@
+pub mod commitment_snapshot;
+pub mod ema_snapshot;
+pub mod epoch_snapshot;
+
+pub use commitment_snapshot::*;
+pub use ema_snapshot::*;
+pub use epoch_snapshot::*;
